requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0
aiofiles==23.2.1
Pillow==10.3.0
python-dotenv==1.0.1
# serpapi==0.1.5  ← ya no se usa
thefuzz==0.22.1
python-Levenshtein==0.25.1
cachetools==5.3.3

# ── Modelo ViT (clasificador de Pokémon) ──────────────────────────
transformers==4.41.0
torch==2.3.0
torchvision==0.18.0

# ── Gemini Flash (fallback gratuito cuando ViT < 80%) ────────────────
google-generativeai==0.7.2